Introduce ReplParser and use it in es.py.
Now, when you hit enter, if you've typed something that isn't a complete
Script, the parser state is set aside. You're prompted for another line
of input, and that's fed to the parser next, until the Script is
complete or there's a syntax error.

```
> x =
» y +
» z;
('ExpressionStatement', ('AssignmentExpression 4', 'x', '=', ('AdditiveExpression 1', 'y', '+', 'z')), ';')
```

(The `»` is the secondary prompt.)

In theory, this is linear where SpiderMonkey's REPL is quadratic in the
input size, so it should already be a speed win, when pasting huge
amounts of code into the shell. ;-)

The behavior is weird in a lot of cases, for now, because of bugs in the
parser. In particular, ASI as implemented can insert *any* semicolon,
including the EmptyStatement semicolon, which is wrong and causes
`while (true)` to parse as complete when it shouldn't.