Fix ASI bug: infinite EmptyStatements.
The bug involved erroneous input like `{`—literally a program consisting
of a single curly brace.  `JSLexer` would insert a semicolon at the end,
where the parser would interpret it as an EmptyStatement; but that would
leave the lexer in the same state as before, so it would then happen
again forever.

The spec says that ASI does not insert EmptyStatements, so there's a
pretty easy fix.