Type inference for reduce expressions.
Class Grammar has a new attribute, .nt_types, which maps nonterminal
names (always strings, never Apply objects) to types.

The Grammar.method attribute now maps method names to MethodType
objects, which give the argument types and return type, not just the
number of arguments.

The new file, jsparagus/types.py, does Hindley-Milner type
inference. This might be overkill, but let's see if it gets decent
results on the JS grammar. The point of this is to support better Rust
codegen, without requiring us to annotate every single nonterminal and
production with a type.  The better Rust codegen isn't done in this
patch, just the type-checking.