Be more careful with default reduce actions.
To ensure that the reduce action being generated here will actually work
at run time, the number of concrete elements in the rhs must be exactly
1. Before the patch, we were only checking the number of elements, not
the number of concrete elements.

As it turns out, it didn't matter; it was impossible to trigger an
actual bug here. But it was unnecessarily precarious.