Fix bug in State.traceback() that affected comments in output.
The Python output previously said something like this:

    # 603. "for" [looakhead != "let"] "var" VariableDeclarationList[~In, ~Yield, ~Await]

The traceback of States was correctly built, but in turning that into a
string, State.traceback() mistakenly treated one of the States as having
been reached by matching a lookahead restriction. Of course those don't
actually exist in source; the result of State.traceback() should contain
only terminals and nonterminals (including Apply nonterminals).

With this patch, the output is:

    # 603. "for" "(" "var" VariableDeclarationList[~In, ~Yield, ~Await]