Better type names from type inference.
In type inference, when unifying two type variables (with the same
precedence), prior to this change we went with the name of the "actual"
type over the "expected" type, meaning if we had a production

    Statement ::= IfStatement      => $1

then the type name might end up being `IfStatement`, but never
`Statement`. This patch simply reverses this rule, so we prefer the
"expected" type (`Statement`).

With this patch, the type of expressions (using the ES grammar) is
`Expression` and the type of statements is `ModuleItem`.