Convert all names of nonterminals in Productions to Nt objects.
Previously, we never created Nt objects with empty arguments; now that
is normal and all non-parameterized nonterminals are represented that
way everywhere they appear in a production right-hand side.

This also converts all init_nts to Nt objects.

The keys of grammar.nonterminal become Nt objects in
expand_function_nonterminals.

The order of some reductions in the output changes because
expand_function_nonterminals was partly reimplemented and the new
implementation does not retain the order of the input. The new order and
the old order are both deterministic, but they're slightly different
algorithms and produce the results in a different order.