Add Grammar.intern() method.
Interning is making all equal values be the same object. Intern
consistently and we could delete all the __hash__ and __eq__ methods and
use the default ones, which use pointer identity.

This commit doesn't go far enough to get that result, but I did some of
the work and it seems worth keeping.