Stop trying to cope with parameterized nonterminals in JS for now.
It takes too long to generate and compile the parser; the resulting
parser is too big; we almost certainly don't want to handle it this way
in the long run anyway.

Instead, the lexer should be able to ask "Is the Yield bit set?" at the
point where it has actually encountered `yield` and the parser needs to
be able to answer that question. It needs to be a runtime bit, rather
than part of the state.

This cuts parser_generated.rs to 14% of its size (in bytes).