Fix can_close, which was a bit broken by error recovery.
The bug: can_close() was returning false after `2 + 2`, failing to do
ASI at end of input.

I had a very ugly hack where the lexer was monkeypatching the builder
with a handle_error method, in order to make ASI work -- the most
despicable nonsense.  The hack interacted badly with another hack:
Parser.can_close works by cloning the parser. Since the ugly hack is
part of the *lexer*, which is *not* cloned, ASI didn't work during
can_close.

The fix is extremely satisfactory. The hack is removed, and the new
replacement for handle_error is where it belongs: JSParser.on_recover.