Preliminary support for actions in grammars. Untested.
The idea is that after any production in an .esgrammar or .pgen file,
write `=> methodname($0, $1, $3)` to specify what method is called when
that production matches and which arguments are passed.

The numbers refer to grammar symbols in the production; lookahead rules
and ErrorToken don't count since they don't match anything in the text
being parsed.

`Some($0)` and `None` are also supported, but that's all. I know this
isn't enough, but it's a start, it was very easy to do, and it keeps
both Python and Rust working.