Rename one meaning of "action" to "reducer", part 1.
The goal is to only have one thing called "action".

Since any standard treatment refers to the "action table", we'll keep
that use of the word "action". Reduce expressions should not also be
called "action", so we'll switch to using "reduce expression"
everywhere, or "reducer" for short.

This first part of three merely renames a few grammatical productions.