Use an error Result in `for await`. Closes #29.
There are five remaining `panic!()` calls in AstBuilder, and they're how
I want them. Two are in cases that are impossible to reach, due to local
logic that's pretty easy to follow. The other three happen if you pass
an invalid argument to a method, and the ES grammar guarantees that the
parser never will.  They're like `assert!(false)`.