Improvements to handling of Identifier and IdentifierName.
Many more conditional keywords now work; getters, setters, and static
methods work.

The new concept is "synthetic terminals", of which we have two in
ES, *Identifier* and *IdentifierName*. These are defined as sets of
nonsynthetic terminals. At run time, synthetic terminals don't
exist. The lexer produces only nonsynthetic terminals, so those are the
ones that appear in the parser tables. Synthetic terminals are used
purely to make sense of the ES grammar.

There may be some bugs in this new idea. State.traceback() should be
improved.