Factor out functions to load the ES grammars.
Before this patch, we had two modules: parse_esgrammar for reading the
file, and generate_js_parser_tables which did a specific task.

Now there's a module for reading the grammar file that knows the few
details about ECMAScript needed to make an accurate Grammar object,
in addition to what's in the file.