Types: Assign all terminals the type `Token`.
Previously, we assigned constant terminals UnitType, on the theory that
we could optimize such values away instead of pushing them to the
stack. But there are some problems with that.

First, shift actions would have to be more complex--right now they
literally just push the current token to the stack.

Second, there was a bug in the existing code: it should have counted
synthetic terminals as non-constant.

Third, in ES, keyword tokens can be used as Identifiers, in which case
obviously it would be a bug to optimize them away entirely.

Fourth, we could end up needing secondary information from tokens that
we optimize away, like location information.

And anyway, the right way to optimize away stack entries is based on
whether or not they're used, not by pretending they're some other type.

For all these reasons, every terminal symbol already had the type
`Token` in the generated Rust code, so this change just aligns the
parser generator with the pre-existing ground truth.