Add 'Name' to list of terminals when loading the ES grammar.
`Name` is not in the spec. It's what our lexer produces for identifier
tokens that aren't keywords, conditional or otherwise. (The way this
works in the spec is not directly implementable.) This patch tries to
make clear which nonterminals are straight out of the spec and which
ones are our own practical hacks.