Stop computing variable_terminals from the lexical grammar.
Before this change, when loading the syntactic grammar, we used the
lexical grammar to obtain the list of lexical nonterminals, all of which
were treated as terminals in syntactic productions.

This resulted in a `variable_terminals` set that includes all sorts of
constituents of tokens, like *ExponentIndicator* and *Hex4Digits*.

For now, we use a fixed list. This is inflexible, but allows us to load
the syntactic grammar without having to care about the lexical grammar.