Use dataclasses to replace a few more namedtuples.
Dataclasses don't "unpack" (in destructuring assignment) like tuples, so
a few uses of Edge and one use of APS are written more explicitly.  The
effect is not bad; the APS case is even an improvement.

In one place we use `sorted()` on a list of `LRItem`s, so `class LRItem`
has to have `@dataclass(order=True)`.