Method names should be strings.
(At this point in the code, it seems like `nt` could be an Nt object,
and mypy agrees. In practice, either it never actually is an Nt, or
it works out OK regardless because we always end up converting it to a
str on output. This change makes the code more obviously-correct.)