Refine type annotations on ordered set types.
The parser generator contains some uses of, for example, `x | y`, where
x is an OrderedSet and y is an OrderedFrozenSet. To support this, the
type annotation on the `__or__` method needs to take an `AbstractSet`,
Python's generic abstract base class for sets.

In order to conform to AbstractSet, __contains__ methods must accept any
object, regardless of the type parameter.