Distinguish Action.update_stack from Action.follow_edge.
The 2 actions used to be joined but will no longer remain joined. Thus,
`update_stack` is used to express how the parser stack should be mutated, and
`follow_edge` expresses whether the epsilon edge on which the action is attached
should be followed or not. Reduce actions are not following the epsilon edge,
whereas Unwind action would follow the epsilon edge resulting in a split of
concepts between the ordinary stack automaton and a regular automaton managing
the parser stack.

 - Add Unwind action.
 - Change Reduce action to wrap the Unwind action with no follow_edge.
 - Add StackDiff to express parser stack mutations.
 - Add Action.update_stack_with to return a StackDiff.