Update APS.shift_next to reduce by replaying the non-terminal.
This matches the way Reduce actions are implemented in the Python and Rust
runtime. However, this requires an extra `reducing` flag to prevent infinite
loops of APS when matching empty productions with inconsistent parse tables.

This change is useful for the upcoming Replay action, as we would have to split
the action state which is being visited from the stack and shift states.

This change also fix the test case added in #468, by using a reducing path which
is not capable of shifting the reduced non-terminal. By doing so we allow
shifting epsilon edges which are not updating the stack, thus discovering
lookahead which used to be ignored before.