Parse Table: Add stable state numbering. (#465)
Parse Table: Add stable state numbering for human readable output.

When doing grammar changes or parse table transformations, the state
might be renumbered. While this is a useful thing for optimization, this
makes thing terribly hard for debugging.

This change introduces a `stable_hash` to each state, which relies on the
same content used to give a unique hash identifier for states. Edge, APS,
Actions and states now have an additional `stable_str`, which behaves like
`__str__` except that it expect a list of `states` coming from the parse table,
to provide a stable output.

In verbose mode, the parse table is dump after each phase such that it can
easily be compared after doing modifications of the grammar.