Rust backend: Convert action match to a table of functions.
This patch changes the big match over the set of actions, to be a table indexed
by the state. This change the generated code and adds the ability to inline the
action function, thus avoiding the overhead of the function call.

However, due to the fact that the `Handler` type is generic, we have to
wrap the constant table definition in an `impl` block to specialize it based on
the caller specialization.