Rust/Python backends: Add action state arguments.
This change is made to delay when replyed term-value are added to the stack. The
intent being that if we delay the replayed term, we might be able to avoid
popping it from the stack when reducing.

This change adds to each state an arguments field which count the number of
arguments which are expected to be replayed by previous states.