Support for FunCall/Unwind with negative offset/replay.
This change is needed in order to convert terms given as argument to be popped
while not being encoded on the stack but on the list of arguments of the state
action.

Having a negative `offset` for `Funcall` or a negative `replay` for Unwind is
equivalent to shifting the list of terms which have to be replayed with dummy
state values, before popping them out of the stack with `Unwind` action.

This change is minimal in the `APS`, because most of the logic is fortunately
already handled in `reduce_path` which computes the `depth` as `pop + replay`
which works as expected in cases where the `replay` value is negative.