Fold Replay followed by Unwind action using arguments.
This change adds the ability to use action state arguments and transfer it as a
popped argument of the following reduce/unwind action.

When unwinding/reducing with arguments, the argument is used as if it were the
next replayed element. Thus we can remove unnecessary replay calls of the
returned values of the AstBuilder.