Refactor scope handling in non-local jump to support lexical scope with environment object
When leaving lexical scope, we need to emit different opcode depending on the
lexical scope (whether it has EnvironmentObject or not).

To provide the information when entering scope hole, EmitterScopeStack has to
return an iterator to walk up the stack, that yields what kind of scope it is.

EmitterScopeWalker yields a outer/inner pair of EmitterScope, while walking up
the scope stack.
NonLocalExitControl iterates over it and calls appropriate method depending on
the scope type.