Skip walking back epsilon edges on consistent parse tables.
When the parse table is consistent, the epsilon edges are unconditionally taken
and executed. Thus the valid reasoning that we should traverse all the backedges
to find states hidden behind epsilon edges which might reduce a non-terminals,
or that we might have a non-terminal shift next to an epsilon action does not
hold.

This optimization is nice to have, as it reduces the amount of iteration needed.
This also avoid removing an assertion which verifies that the first state of the
path is capable of shifting the reduced non-terminal.