Add smoosh_tools bench command.
The bench command will look for a a checkout of real-js-samples and use it for
comparing both SpiderMonkey and SmooshMonkey parsers.

It relies on the benchmarks/compare-spidermonkey-parsers.js file as a harness to
run both parsers on a all files contained in a directory.

The directory is by default set to the real-js-samples, or it can be overriden
by using the --samples-dir=<DIR> command line option.