Optimize with filter and replay (#578)
* Add lower_reduce_actions transformation.

This transformation convert Reduce(Unwind(...)) actions into its Unwind action
followed by FilterStates and Replay edges.

This transformation pre-compute the determinisitic path taken while reducing,
and adds it back to the graph. Before this transformation, the parser stack was
completely controlled by shift-ing edges and reducing-non-terminals. After this
transformation, the state of the Parser might be disconnected from the last
state pushed to the stack.

* Add fold_by_destination while rewriting state indexes.

When rewriting state indexes, such as in fold_identical_endings, some edges
might be rewritten to edges which already exists.

For example, 2 FilterStates conditions, with different state indexes, might be
rewritten to the same condition with identical destinations. In such cases, the
rewritten state indexes would cause th state to become inconsistent, as
FilterStates conditions are consistent iff there is no overlap of state indexes
between the conditions.

Adding fold_by_destination solves this problem by folding conditions which have
the same destination, thus removing the problem as the condition would be merged
after rewriting the state indexes.

* Fix typo in Replay slots name.

* Update term_is_shifted to considere follow_edge status of Actions.

The function `term_is_shifted` considers whether an edge is followed or not, as
opposed to `term_is_stacked` which considers whether the term of the edge is
added to the parser stack when the edge is followed.

* Skip walking back epsilon edges on consistent parse tables.

When the parse table is consistent, the epsilon edges are unconditionally taken
and executed. Thus the valid reasoning that we should traverse all the backedges
to find states hidden behind epsilon edges which might reduce a non-terminals,
or that we might have a non-terminal shift next to an epsilon action does not
hold.

This optimization is nice to have, as it reduces the amount of iteration needed.
This also avoid removing an assertion which verifies that the first state of the
path is capable of shifting the reduced non-terminal.

* Update parse_pgen.

This follow the remove_reduce optimization as well as the grouping on states
based on their successors.