Fold actions by destination using the ParseTable.
Previously, when rewriting state indexes, we would fold actions together.
However, this breaks an invariant which is that each edge appears in the
backedge of its destinations.

By folding under `StateAndTransitions`, we do not have access to the destination
state, we only have the `StateId`.

By moving the code to the `ParseTable`, we can rely on `remove_edge` and
`add_edge` to keep these invariant in place.