Add ScopeKind::FunctionParametersAndBody to handle unsupported cases properly
Unsupported binding identifier at top-level script is ignored because
there's no ScopeKind on the stack.

The same thing should happen for functions.
`ScopeKind::FunctionName` is on to the stack while visiting the function, to
handle function name.

Push `ScopeKind::FunctionParametersAndBody` while visiting the function
parameters and body, in order to disable the effect of
`ScopeKind::FunctionName`.