Fix folding causing inconsistencies (#606)
* Fold actions by destination using the ParseTable.

Previously, when rewriting state indexes, we would fold actions together.
However, this breaks an invariant which is that each edge appears in the
backedge of its destinations.

By folding under `StateAndTransitions`, we do not have access to the destination
state, we only have the `StateId`.

By moving the code to the `ParseTable`, we can rely on `remove_edge` and
`add_edge` to keep these invariant in place.

* Remove assertion from `ParseTable.remove_backedge`.

remove_backedge is an internal function which is not balanced as the edge being
removed should also be removed from the forward edges.

However this condition does not hold under `clear_edges` which remove all
outgoing edges at once.

* Add an assertion to check all state invariants of the ParseTable.

* Prevent folding which might cause bogus state.

When folding 2 states, for example state 218 and 219, we would have to rewrite
the state indexes of actions which are referencing these state. However, we
might have a single state, which has a condition for both of these states with a
destination which may be different for each of these states, such as:

3444:
  FilterStates(OrederedSet([218])) --> 3048
  FilterStates(OrederedSet([219])) --> 3049

Folding 218 in 219 would cause the state 3444 to have 2 edges with the same term
and 2 different destination, which is bogus.

The code of this patch fixes this issue by making collecting all the states
which are being referenced, as back-reference, indexed by the state which are
being referenced. Then it ensure that if we ignore the destination then we have
the same number of unique source & terms.

Thus, using this trick we would prevent this folding from happening until the
states 2048 and 2049 are not folded, which ought to happen at some later
iteration of `fold_identical_endings`.

* Document why we ignore the content of maybe_unreachable_set in rewrite_state_indexes.