Add state arguments for replay (#608)
* Rust/Python backends: Add action state arguments.

This change is made to delay when replyed term-value are added to the stack. The
intent being that if we delay the replayed term, we might be able to avoid
popping it from the stack when reducing.

This change adds to each state an arguments field which count the number of
arguments which are expected to be replayed by previous states.

* Support for FunCall/Unwind with negative offset/replay.

This change is needed in order to convert terms given as argument to be popped
while not being encoded on the stack but on the list of arguments of the state
action.

Having a negative `offset` for `Funcall` or a negative `replay` for Unwind is
equivalent to shifting the list of terms which have to be replayed with dummy
state values, before popping them out of the stack with `Unwind` action.

This change is minimal in the `APS`, because most of the logic is fortunately
already handled in `reduce_path` which computes the `depth` as `pop + replay`
which works as expected in cases where the `replay` value is negative.

* Fold Replay followed by Unwind action using arguments.

This change adds the ability to use action state arguments and transfer it as a
popped argument of the following reduce/unwind action.

When unwinding/reducing with arguments, the argument is used as if it were the
next replayed element. Thus we can remove unnecessary replay calls of the
returned values of the AstBuilder.

* Unwind can now have a negative replay value.

* Update parse_pgen.