Add optional skip-list argument for profiling.
As SmooshMonkey is still being developped, not all tests are parsed succesfully
with SmooshMonkey. Thus the results cannot be compared using a profile as the
profile might contain extra samples of one parser, and error path of the other.

This change adds a skip-list to compare-spidermonkey-parsers.js benchmark, such
that profiles can safely be used to compare the ratio of each parts.

Note: When the skip-list is not representative of skipped file, it is being
updated accordingly. Which implies that one should run the benchmark twice, once
to fill the skip-list, and a second time to collect the profile.