Add a new way to fix inconsistencies by using existing conditionals.
When we have conditionals which are checking whether the previous character is
on a new line or not, such as `[no LineTerminator here]`, it is safe to wrap any
non-garded edge with both the condition and its negated version.

This way, the resolution of lookahead or reduce-reduce and shift-reduce
conflict, does not have to push the check for new lines further than necessary.