Simplify fix_inconsistent_table.
This change adds `mark_sweep_states` function to replace unreachable states by
`None`, and prevent them from being reused. This removes the need for tracking
inconsistent states which are potentially unreachable, but we are to re-scan the
whole parse table in order to refill the todo list in cased of missed
inconsistencies.