Prevent aliasing of Reduce actions.
Lower Reduce Action lookup for the state which would be used to shift the
non-terminal, and then encode the Replay action based on the the state where we
are reducing the nonterminal.

However, this strategy alone does not work when reduce actions are shared
across, as having different lookahead with the same reduce actions, we generate
multiple Replay paths for the same state which shift the reduced nonterminal.

Adding the Lookahead terms to the reduce action, ensure that we would not add
sharing if we did not have sharing in LR0 before.